created Contact.java for inheritance
